requirements.txt
jsonschema==4.15.0
Pillow==9.2.0
requests==2.28.1
streamlit==1.12.2
